+++ changelog.md
@@ -4,6 +4,8 @@
 
 # Unreleased
 
+# 3.1.7
+
 - Improve price modifier
 - Fix confirm button with commands and message, you need to use now a ``confirm-actions``